Add draw grid with custom number of squares functionality
The grid can now be redrawn by changing the content of the number input
below it. Each time the contents of the grid are flushed and it gets
filled with new squares with the correct height and width so that
the number of squares per side chosen is respected and the grid is
filled.